Incorporated comments, and finished refactoring.
