+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-client-dto/src/main/java/org/apache/nifi/web/api/dto/status/PortStatusDTO.java
@@ -32,7 +32,6 @@ public class PortStatusDTO {
     private Boolean transmitting;
     private String runStatus;
     private Date statsLastRefreshed;
-    private Boolean allowRemoteAccess;
 
     private PortStatusSnapshotDTO aggregateSnapshot;
     private List<NodePortStatusSnapshotDTO> nodeSnapshots;
@@ -124,19 +123,5 @@ public class PortStatusDTO {
 
     public void setStatsLastRefreshed(Date statsLastRefreshed) {
         this.statsLastRefreshed = statsLastRefreshed;
-    }
-
-    /**
-     * @return whether this port can be accessed remotely via Site-to-Site protocol.
-     */
-    @ApiModelProperty(
-        value = "Whether this port can be accessed remotely via Site-to-Site protocol."
-    )
-    public Boolean isAllowRemoteAccess() {
-        return allowRemoteAccess;
-    }
-
-    public void setAllowRemoteAccess(Boolean allowRemoteAccess) {
-        this.allowRemoteAccess = allowRemoteAccess;
     }
 }
