+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/java/org/apache/nifi/reporting/StandardEventAccess.java
@@ -408,7 +408,6 @@ public class StandardEventAccess implements UserAwareEventAccess {
             // special handling for public ports
             if (port instanceof PublicPort) {
                 portStatus.setTransmitting(((PublicPort) port).isTransmitting());
-                portStatus.setAllowRemoteAccess(true);
             }
 
             final FlowFileEvent entry = statusReport.getReportEntries().get(port.getIdentifier());
