+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/resources/FlowConfiguration.xsd
@@ -157,8 +157,8 @@
             <!-- Each "processor" defines the actual dataflow work horses that make dataflow happen-->
             <xs:element name="processor" type="ProcessorType" minOccurs="0" maxOccurs="unbounded"/>
 
-            <xs:element name="inputPort" type="LocalPortType" minOccurs="0" maxOccurs="unbounded"/>
-            <xs:element name="outputPort" type="LocalPortType" minOccurs="0" maxOccurs="unbounded"/>
+            <xs:element name="inputPort" type="PublicPortType" minOccurs="0" maxOccurs="unbounded"/>
+            <xs:element name="outputPort" type="PublicPortType" minOccurs="0" maxOccurs="unbounded"/>
 
             <xs:element name="label" type="LabelType" minOccurs="0" maxOccurs="unbounded" />
             <xs:element name="funnel" type="FunnelType" minOccurs="0" maxOccurs="unbounded" />
@@ -188,8 +188,9 @@
         </xs:sequence>
     </xs:complexType>
 
-    <!-- Same as ProcessGroupType except that instead of input ports & output ports being of type PortType,
-    they are of type RootGroupPortType -->
+    <!-- Same as ProcessGroupType except:
+         - RootProcessGroupType doesn't have versionControlInformation
+         - RootProcessGroupType doesn't have variable -->
     <xs:complexType name="RootProcessGroupType">
         <xs:sequence>
             <xs:element name="id" type="NonEmptyStringType" />
@@ -201,8 +202,8 @@
             <!-- Each "processor" defines the actual dataflow work horses that make dataflow happen-->
             <xs:element name="processor" type="ProcessorType" minOccurs="0" maxOccurs="unbounded"/>
 
-            <xs:element name="inputPort" type="RootGroupPortType" minOccurs="0" maxOccurs="unbounded"/>
-            <xs:element name="outputPort" type="RootGroupPortType" minOccurs="0" maxOccurs="unbounded"/>
+            <xs:element name="inputPort" type="PublicPortType" minOccurs="0" maxOccurs="unbounded"/>
+            <xs:element name="outputPort" type="PublicPortType" minOccurs="0" maxOccurs="unbounded"/>
 
             <xs:element name="label" type="LabelType" minOccurs="0" maxOccurs="unbounded" />
             <xs:element name="funnel" type="FunnelType" minOccurs="0" maxOccurs="unbounded" />
@@ -311,24 +312,13 @@
         </xs:sequence>
     </xs:complexType>
 
-    <xs:complexType name="RootGroupPortType">
+    <xs:complexType name="PublicPortType">
         <xs:complexContent>
             <xs:extension base="PortType">
                 <xs:sequence>
-                    <xs:element name="maxConcurrentTasks" type="xs:positiveInteger"></xs:element>
+                    <xs:element name="maxConcurrentTasks" type="xs:positiveInteger" minOccurs="0" ></xs:element>
                     <xs:element name="userAccessControl" type="xs:string" minOccurs="0" maxOccurs="unbounded" />
                     <xs:element name="groupAccessControl" type="xs:string" minOccurs="0" maxOccurs="unbounded" />
-                    <xs:element name="allowRemoteAccess" type="xs:boolean" minOccurs="0" maxOccurs="1" />
-                </xs:sequence>
-            </xs:extension>
-        </xs:complexContent>
-    </xs:complexType>
-
-    <xs:complexType name="LocalPortType">
-        <xs:complexContent>
-            <xs:extension base="PortType">
-                <xs:sequence>
-                    <xs:element name="maxConcurrentTasks" type="xs:positiveInteger"></xs:element>
                     <xs:element name="allowRemoteAccess" type="xs:boolean" minOccurs="0" maxOccurs="1" />
                 </xs:sequence>
             </xs:extension>
