+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/authorization/StandardAuthorizableLookup.java
@@ -231,7 +231,7 @@ class StandardAuthorizableLookup implements AuthorizableLookup {
     }
 
     @Override
-    public PublicPortAuthorizable getRootGroupOutputPort(String id) {
+    public PublicPortAuthorizable getPublicOutputPort(String id) {
         final Port outputPort = outputPortDAO.getPort(id);
 
         if (!(outputPort instanceof PublicPort)) {
