+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/WEB-INF/partials/canvas/new-port-dialog.jsp
@@ -23,15 +23,13 @@
                 <input id="new-port-name" type="text"/>
             </div>
         </div>
-        <div class="setting">
-            <div class="remote-access-setting">
-                <div class="setting-name">
-                    <span id="port-allow-remote-access-label"></span>
-                    <div id="port-allow-remote-access-info" class="fa fa-question-circle" alt="Info"></div>
-                </div>
-                <div class="setting-field">
-                    <div id="port-allow-remote-access"></div>
-                </div>
+        <div id="port-allow-remote-access-setting" class="setting">
+            <div class="setting-name">
+                <span id="port-allow-remote-access-label"></span>
+                <div id="port-allow-remote-access-info" class="fa fa-question-circle" alt="Info"></div>
+            </div>
+            <div class="setting-field">
+                <div id="port-allow-remote-access"></div>
             </div>
         </div>
     </div>
