+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/js/nf/canvas/nf-policy-management.js
@@ -1651,10 +1651,10 @@
                 $('#component-policy-target')
                     .combo('setOptionEnabled', {
                         value: 'write-receive-data'
-                    }, nfCanvasUtils.isInputPort(selection) && d.component.allowRemoteAccess === true)
+                    }, nfCanvasUtils.isInputPort(selection) && d.allowRemoteAccess === true)
                     .combo('setOptionEnabled', {
                         value: 'write-send-data'
-                    }, nfCanvasUtils.isOutputPort(selection) && d.component.allowRemoteAccess === true)
+                    }, nfCanvasUtils.isOutputPort(selection) && d.allowRemoteAccess === true)
                     .combo('setOptionEnabled', {
                         value: 'read-data'
                     }, !nfCanvasUtils.isLabel(selection))
