+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/js/nf/canvas/nf-port-details.js
@@ -77,7 +77,7 @@
                 // populate the port settings
                 nfCommon.populateField('read-only-port-name', selectionData.component.name);
                 nfCommon.populateField('read-only-port-id', selectionData.id);
-                nfCommon.populateField('read-only-port-allow-remote-access', true === selectionData.component.allowRemoteAccess ? 'true' : 'false');
+                nfCommon.populateField('read-only-port-allow-remote-access', true === selectionData.allowRemoteAccess ? 'true' : 'false');
                 nfCommon.populateField('read-only-port-concurrent-tasks', selectionData.component.concurrentlySchedulableTaskCount);
                 nfCommon.populateField('read-only-port-comments', selectionData.component.comments);
